fix downstream crates

+++ pallas-applying/src/alonzo.rs
@@ -378,7 +378,7 @@ fn check_tx_ex_units(mtx: &MintedTx, prot_pps: &AlonzoProtParams) -> ValidationR
         match &tx_wits.redeemer {
             Some(redeemers_vec) => {
                 let mut steps: u64 = 0;
-                let mut mem: u32 = 0;
+                let mut mem: u64 = 0;
                 for Redeemer { ex_units, .. } in redeemers_vec {
                     mem += ex_units.mem;
                     steps += ex_units.steps;
